+++ lib/storageHandler.dart
@@ -1,0 +1,44 @@
+import 'dart:convert';
+import 'dataset.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'dart:async';
+import 'dart:io';
+
+class StorageHandler {
+
+
+Future<Dataset> loadDatasetFromStorage(String name) async {
+    final prefs = await SharedPreferences.getInstance();
+    String jsonString = prefs.getString(name);
+    return Dataset.fromJson(jsonString);
+  }
+
+  List<String> jsonifyImages(List<File> images) {
+    List<String> imagesInString;
+    for (File file in images) {
+      imagesInString.add(file.path);
+    }
+    return imagesInString;
+  }
+
+  Future<void> saveDataset(Dataset dataset) async {
+    Map<String, dynamic> jsonMap = {
+      'name': dataset.name,
+      'rightSwipeName': dataset.rightSwipeName,
+      'rightSwipeTag': dataset.rightSwipeTag,
+      'leftSwipeName': dataset.leftSwipeName,
+      'leftSwipeTag': dataset.leftSwipeTag,
+      'images': jsonifyImages(dataset.images)
+    };
+
+    final prefs = await SharedPreferences.getInstance();
+    prefs.setString(dataset.name, jsonify(jsonMap));
+    prefs.setString("activeDataset", dataset.name);
+  }
+
+  String jsonify(Map<String, dynamic> toJson) {
+    return jsonEncode(toJson);
+  }
+
+
+}
